requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
alabaster=0.7.12=pyhd3eb1b0_0
altgraph=0.17.2=pypi_0
appdirs=1.4.4=pyhd3eb1b0_0
argh=0.26.2=py38_0
arrow=0.13.1=py38_0
astroid=2.6.6=py38h06a4308_0
async_generator=1.10=pyhd3eb1b0_0
atomicwrites=1.4.0=py_0
attrs=21.2.0=pyhd3eb1b0_0
autopep8=1.5.7=pyhd3eb1b0_0
babel=2.9.1=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
binaryornot=0.4.4=pyhd3eb1b0_1
black=19.10b0=py_0
blas=1.0=mkl
bleach=4.0.0=pyhd3eb1b0_0
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.10.26=h06a4308_2
certifi=2021.10.8=py38h06a4308_0
cffi=1.14.6=py38h400218f_0
chardet=4.0.0=py38h06a4308_1003
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.3=pyhd3eb1b0_0
cloudpickle=2.0.0=pyhd3eb1b0_0
colorama=0.4.4=pyhd3eb1b0_0
cookiecutter=1.7.2=pyhd3eb1b0_0
cpuonly=2.0=0
cryptography=35.0.0=py38hd23ed53_0
dbus=1.13.18=hb2f20db_0
debugpy=1.5.1=py38h295c915_0
decorator=5.1.0=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
diff-match-patch=20200713=pyhd3eb1b0_0
docutils=0.17.1=py38h06a4308_1
entrypoints=0.3=py38_0
expat=2.4.1=h2531618_2
ffmpeg=4.3=hf484d3e_0
flake8=3.9.2=pyhd3eb1b0_0
fontconfig=2.13.1=h6c09931_0
freetype=2.11.0=h70c0345_0
giflib=5.2.1=h7b6447c_0
glib=2.69.1=h5202010_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
icu=58.2=he6710b0_3
idna=3.2=pyhd3eb1b0_0
imagesize=1.2.0=pyhd3eb1b0_0
importlib-metadata=4.8.1=py38h06a4308_0
importlib_metadata=4.8.1=hd3eb1b0_0
inflection=0.5.1=py38h06a4308_0
intel-openmp=2021.4.0=h06a4308_3561
intervaltree=3.1.0=pyhd3eb1b0_0
ipykernel=6.4.1=py38h06a4308_1
ipython=7.29.0=py38hb070fc8_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
isort=5.9.3=pyhd3eb1b0_0
jedi=0.18.0=py38h06a4308_1
jeepney=0.7.1=pyhd3eb1b0_0
jinja2=2.11.3=pyhd3eb1b0_0
jinja2-time=0.2.0=pyhd3eb1b0_2
jpeg=9d=h7f8727e_0
jsonschema=3.2.0=pyhd3eb1b0_2
jupyter_client=6.1.12=pyhd3eb1b0_0
jupyter_core=4.9.1=py38h06a4308_0
jupyterlab_pygments=0.1.2=py_0
kaggle-environments=1.8.12=pypi_0
keyring=23.1.0=py38h06a4308_0
lame=3.100=h7b6447c_0
lazy-object-proxy=1.6.0=py38h27cfd23_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h7b6447c_0
libspatialindex=1.9.3=h2531618_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h7f8727e_2
libuv=1.40.0=h7b6447c_0
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
lz4-c=1.9.3=h295c915_1
markupsafe=1.1.1=py38h7b6447c_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mccabe=0.6.1=py38_1
mistune=0.8.4=py38h7b6447c_1000
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
mypy_extensions=0.4.3=py38_0
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.1.0=py38h06a4308_0
nbformat=5.1.3=pyhd3eb1b0_0
ncurses=6.3=h7f8727e_2
nest-asyncio=1.5.1=pyhd3eb1b0_0
nettle=3.7.3=hbbd107a_1
numpy=1.21.4=pypi_0
numpy-base=1.21.2=py38h79a1101_0
numpydoc=1.1.0=pyhd3eb1b0_1
olefile=0.46=pyhd3eb1b0_0
openh264=2.1.0=hd408876_0
openssl=1.1.1l=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pandocfilters=1.4.3=py38h06a4308_1
parso=0.8.2=pyhd3eb1b0_0
pathspec=0.7.0=py_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.4.0=pypi_0
pip=21.2.4=py38h06a4308_0
pluggy=1.0.0=py38h06a4308_0
poyo=0.5.0=pyhd3eb1b0_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
psutil=5.8.0=py38h27cfd23_1
ptyprocess=0.7.0=pyhd3eb1b0_2
pycodestyle=2.7.0=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pydocstyle=6.1.1=pyhd3eb1b0_0
pyflakes=2.3.1=pyhd3eb1b0_0
pygame=2.1.0=pypi_0
pygments=2.10.0=pyhd3eb1b0_0
pyinstaller=4.7=pypi_0
pyinstaller-hooks-contrib=2021.4=pypi_0
pylint=2.9.6=py38h06a4308_1
pyls-spyder=0.4.0=pyhd3eb1b0_0
pyopenssl=21.0.0=pyhd3eb1b0_1
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py38h05f1152_4
pyrsistent=0.18.0=py38heee7806_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.12=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-lsp-black=1.0.0=pyhd3eb1b0_0
python-lsp-jsonrpc=1.0.0=pyhd3eb1b0_0
python-lsp-server=1.2.4=pyhd3eb1b0_0
python-slugify=5.0.2=pyhd3eb1b0_0
pytorch=1.10.0=py3.8_cpu_0
pytorch-mutex=1.0=cpu
pytz=2021.3=pyhd3eb1b0_0
pyxdg=0.27=pyhd3eb1b0_0
pyyaml=6.0=py38h7f8727e_1
pyzmq=22.2.1=py38h295c915_1
qdarkstyle=3.0.2=pyhd3eb1b0_0
qstylizer=0.1.10=pyhd3eb1b0_0
qt=5.9.7=h5867ecd_1
qtawesome=1.0.3=pyhd3eb1b0_0
qtconsole=5.1.1=pyhd3eb1b0_0
qtpy=1.10.0=pyhd3eb1b0_0
readline=8.1=h27cfd23_0
regex=2021.8.3=py38h7f8727e_0
requests=2.26.0=pyhd3eb1b0_0
rope=0.21.1=pyhd3eb1b0_0
rtree=0.9.7=py38h06a4308_1
secretstorage=3.3.1=py38h06a4308_0
setuptools=58.0.4=py38h06a4308_0
sip=4.19.13=py38he6710b0_0
six=1.16.0=pyhd3eb1b0_0
snowballstemmer=2.1.0=pyhd3eb1b0_0
sortedcontainers=2.4.0=pyhd3eb1b0_0
sphinx=4.2.0=pyhd3eb1b0_1
sphinxcontrib-applehelp=1.0.2=pyhd3eb1b0_0
sphinxcontrib-devhelp=1.0.2=pyhd3eb1b0_0
sphinxcontrib-htmlhelp=2.0.0=pyhd3eb1b0_0
sphinxcontrib-jsmath=1.0.1=pyhd3eb1b0_0
sphinxcontrib-qthelp=1.0.3=pyhd3eb1b0_0
sphinxcontrib-serializinghtml=1.1.5=pyhd3eb1b0_0
spyder=5.1.5=py38h06a4308_1
spyder-kernels=2.1.3=py38h06a4308_0
sqlite=3.36.0=hc218d9a_0
testpath=0.5.0=pyhd3eb1b0_0
text-unidecode=1.3=pyhd3eb1b0_0
textdistance=4.2.1=pyhd3eb1b0_0
three-merge=0.1.1=pyhd3eb1b0_0
tinycss=0.4=pyhd3eb1b0_1002
tk=8.6.11=h1ccaba5_0
toml=0.10.2=pyhd3eb1b0_0
torchaudio=0.10.0=py38_cpu
torchvision=0.11.1=py38_cpu
tornado=6.1=py38h27cfd23_0
traitlets=5.1.1=pyhd3eb1b0_0
typed-ast=1.4.3=py38h7f8727e_1
typing_extensions=3.10.0.2=pyh06a4308_0
ujson=4.0.2=py38h2531618_0
unidecode=1.2.0=pyhd3eb1b0_0
urllib3=1.26.7=pyhd3eb1b0_0
watchdog=2.1.3=py38h06a4308_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py38_1
wheel=0.37.0=pyhd3eb1b0_1
whichcraft=0.6.1=pyhd3eb1b0_0
wrapt=1.12.1=py38h7b6447c_1
wurlitzer=2.1.1=py38h06a4308_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
yapf=0.31.0=pyhd3eb1b0_0
zeromq=4.3.4=h2531618_0
zipp=3.6.0=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
